Homework6/requirements.txt
Flask==2.0.3
requests==2.27.1
python-dateutil==2.8.2
Flask-Cors==3.0.10

